seperated frontend common function
